feat: Added images

+++ Jenkinsfile
@@ -89,16 +89,20 @@ pipeline {
                 script {
                     // Install crane if not already available
                     sh '''
-                        curl -LO https://github.com/google/go-containerregistry/releases/download/v${CRANE_VERSION}/crane_amd64_darwin.tar.gz
-                        # Check the downloaded file
-                        file crane_amd64_darwin.tar.gz
-                        tar xzf crane_amd64_darwin.tar.gz
-                        mv crane /usr/local/bin/crane
-                        rm crane_amd64_darwin.tar.gz
+                        if ! command -v crane &> /dev/null
+                        then
+                            echo "crane could not be found, installing..."
+                            curl -LO https://github.com/google/go-containerregistry/releases/download/v${CRANE_VERSION}/crane_amd64_darwin.tar.gz
+                            # Check the downloaded file
+                            file crane_amd64_darwin.tar.gz
+                            tar xzf crane_amd64_darwin.tar.gz
+                            mv crane /usr/local/bin/crane
+                            rm crane_amd64_darwin.tar.gz
+                        fi
                     '''
                     
                     // Push Docker image using crane
-                    sh 'crane cp registry.k8s.io/autoscaling/cluster-autoscaler:v1.29.3 anu398/cluster-autoscaler:v1.29.3'
+                    sh 'crane copy registry.k8s.io/autoscaling/cluster-autoscaler:v1.29.3 anu398/cluster-autoscaler:v1.29.3'
                 }
             }
         }
